requirements.txt
pandas==2.3.3
webdriver-manager==4.0.2
requests-html==0.10.0   
selenium==4.38.0
pyarrow==14.0.1
numpy==1.26.4